requirements.txt
pillow==10.3.0
Send2Trash==1.8.3
PyAutoGUI==0.9.54
pyperclip==1.9.0
